refactor: rename paths module to path
